+++ lectures/week07/12.2-iterators2.cpp
@@ -1,16 +1,17 @@
-#include<iostream>
-using std::cout; using std::endl; 
-#include<vector>
+#include <iostream>
+using std::cout;
+using std::endl;
+#include <vector>
 using std::vector;
-#include<string>
+#include <string>
 using std::string;
-#include<sstream>
+#include <sstream>
 using std::ostringstream;
 
-template<typename T>
-string vec_to_string(const vector<T>& v){
+template <typename T>
+string vec_to_string(const vector<T> &v) {
   ostringstream oss;
-  for(auto iter=v.cbegin(); iter!=v.cend(); ++iter){
+  for (auto iter = v.cbegin(); iter != v.cend(); ++iter) {
     oss << *iter << ", ";
     // *iter = *iter + 1; // can't, it's low-level const
   }
@@ -18,53 +19,49 @@ string vec_to_string(const vector<T>& v){
   return result.substr(0, result.size() - 2);
 }
 
-template<typename T>
-vector<T> rev_vector(vector<T>& v){
+template <typename T>
+vector<T> rev_vector(vector<T> &v) {
   vector<T> result;
-  for(auto iter=v.rbegin(); iter!=v.rend(); ++iter)
+  for (auto iter = v.rbegin(); iter != v.rend(); ++iter)
     result.push_back(*iter);
   return result;
 }
-    
 
-int main (){
-  vector<long> v{1,2,3,4,5,6,7,8,9,10};
-  string s = "abcd1234";  
+int main() {
+  vector<int> v{1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
+  string s = "abcd1234";
 
   cout << "initial:\n" << vec_to_string(v) << endl;
 
   auto rev = rev_vector(v);
-  cout << "reversed vector\n" <<  vec_to_string(rev) << endl;
+  cout << "reversed vector\n" << vec_to_string(rev) << endl;
 
   /* range-for and iter loop equivalents */
-  
+
   // copy, change not reflected
-  for (auto element : v)
-    element += 1;
-  
-  for (auto pos=v.begin(), end=v.end(); pos != end; ++pos){
+  for (auto element : v) element += 1;
+
+  for (auto pos = v.begin(), end = v.end(); pos != end; ++pos) {
     auto element = *pos;
     element += 1;
   }
-  cout <<"first version\n" << vec_to_string(v)<<endl;
+  cout << "first version\n" << vec_to_string(v) << endl;
 
   // reference, change is reflected
-  for (auto &element : v)
-    element += 1;
-  
-  for (auto pos=v.begin(), end=v.end(); pos != end; ++pos){
+  for (auto &element : v) element += 1;
+
+  for (auto pos = v.begin(), end = v.end(); pos != end; ++pos) {
     auto &element = *pos;
     element += 1;
   }
-  cout <<"second version (done twice)\n" << vec_to_string(v)<<endl;  
+  cout << "second version (done twice)\n" << vec_to_string(v) << endl;
 
   // const reference, can't modify
   for (const auto &element : v)
     // element += 1;
-  
-  for (auto pos=v.begin(), end=v.end(); pos != end; ++pos){
-    const auto &element = *pos;
-    // element += 1;
-  }
-  
+
+    for (auto pos = v.begin(), end = v.end(); pos != end; ++pos) {
+      const auto &element = *pos;
+      // element += 1;
+    }
 }
